solidifying knowledge

+++ median.py
@@ -14,9 +14,9 @@ print(chooseMedian([1, 2, 3, 8, 5, 2, 6, 5, 9, 4, 1, 2]))
 def countMedian(numbers):
     numbers = sorted(numbers)
     print(len(numbers))
-    print(numbers[len(numbers) / 2 - 1])
+    print(numbers[int(len(numbers) / 2 - 1)])
     if len(numbers) % 2 == 0:
-        print('Median from this list is a ') + str(((numbers[len(numbers) / 2 - 1]) + (numbers[len(numbers) / 2])) / 2)
+        print('Median from this list is a ') + str(numbers[int(len(numbers) / 2 - 1)] + numbers[int(len(numbers) / 2)] / 2.0)
 
 
 countMedian([1, 2, 3, 5, 7, 4, 3, 4])
